feat: add ImageService interface
New layer of abstraction, encapsulation, separation of concerns
and flexibility.
